+++ src/environments/edc-ui-feature-set.ts
@@ -1,6 +1,0 @@
-/**
- * Active Feature Set.
- *
- * This code base supports both sovity EDC UI and mds EDC UI
- */
-export type EdcUiFeatureSet = 'sovity' | 'mds';