+++ src/environments/environment.ts
@@ -6,7 +6,6 @@ import {EdcUiEnvironment} from "./edc-ui-environment";
 
 export const environment: EdcUiEnvironment = {
   production: false,
-  activeFeatureSet: "sovity"
 };
 
 /*