+++ src/modules/app/config/app-config.service.ts
@@ -1,0 +1,11 @@
+import {AppConfig} from "./app-config";
+import {Injectable} from "@angular/core";
+
+/**
+ * Holds {@link AppConfig}
+ */
+@Injectable()
+export class AppConfigService {
+  // will be set by APP_INITIALIZER
+  public config!: AppConfig;
+}