+++ src/modules/edc-demo/components/asset-detail-dialog/asset-detail-dialog-field-builder.ts
@@ -2,10 +2,13 @@ import {Injectable} from '@angular/core';
 import {Asset} from "../../models/asset";
 import {AssetDetailDialogField} from "./asset-detail-dialog-field";
 import {validUrlPattern} from "../../validators/url-validator";
-import {isFeatureSetActive} from "../../pipes/is-active-feature-set.pipe";
+import {ActiveFeatureSet} from "../../../app/config/active-feature-set";
 
 @Injectable()
 export class AssetDetailDialogFieldBuilder {
+  constructor(private activeFeatureSet: ActiveFeatureSet) {
+  }
+
   buildFields(asset: Asset): AssetDetailDialogField[] {
     const tryValue = (s?: string | null) => s || '-';
     const tryUrl = (s?: string | null): string | undefined => s?.match(validUrlPattern) ? s : undefined
@@ -50,7 +53,7 @@ export class AssetDetailDialogFieldBuilder {
     ];
 
     // MDS Specific Fields
-    if (isFeatureSetActive('mds')) {
+    if (this.activeFeatureSet.isMds()) {
       fields.push(
         {
           icon: 'commute',