+++ src/modules/edc-demo/pipes/is-active-feature-set.pipe.ts
@@ -1,21 +1,17 @@
 import {Pipe, PipeTransform} from "@angular/core";
-import {EdcUiFeatureSet} from "../../../environments/edc-ui-feature-set";
-import {environment} from "../../../environments/environment";
-
-/**
- * Checks whether given feature set is currently active
- * @param featureSet e.g. mds etc.
- */
-export function isFeatureSetActive(featureSet: EdcUiFeatureSet): boolean {
-  return environment.activeFeatureSet === featureSet;
-}
+import {EdcUiFeatureSet} from "../../app/config/edc-ui-feature-set";
+import {ActiveFeatureSet} from "../../app/config/active-feature-set";
 
 /**
  * Use this in angular templates to check active feature set (e.g. "mds" etc.)
  */
 @Pipe({name: 'isActiveFeatureSet'})
 export class IsActiveFeatureSetPipe implements PipeTransform {
+
+  constructor(private activeFeatureSet: ActiveFeatureSet) {
+  }
+
   transform(featureSet: EdcUiFeatureSet): boolean {
-    return isFeatureSetActive(featureSet)
+    return this.activeFeatureSet.is(featureSet)
   }
 }