+++ terraform/output.tf
@@ -1,3 +1,0 @@
-output "dashboard-url" {
-    value = "http://${azurerm_container_group.edc-dashboard.dns_name_label}.${var.location}.azurecontainer.io"
-}